Fix V100 Active Lanes in Warp
Fix for V100 and CUDA 9.0 ballot functions.
Properly use the currently active mask of lanes in a warp instead
of a constant.

We have seen deadlocks in PMacc with when using the `FULL_MASK`
(`0xFFFFFFFF`). This can cause deadlocks e.g. in branched threads
which are properly described with `__activemask()` instead.

https://devblogs.nvidia.com/using-cuda-warp-level-primitives/