fix HIP-clang runtime issues
HIP clang creted code which results into runtime issues.
Some clang optimizations kicking in when the size for the device side
alloc is known at compile time.

Add `volatile` to the size to break optimizations.

note: Tests with CUDA together with PIConGPU did not show negative
effects for the register usage or runtime.