fix wrong alignment
fix #227

In cases where the data allocation size is smaller than the minimal
chunk size the allocation size will be set to minimal chunk size.
In such cases the alignment of the data can be wrong.

mallocMC inerface changes (not visible for the user):
- CreationPolicy::create() get the alignment policy as first template
parameter
- AlignmentPolicy::applyPadding() requires 32bit input sizes and returns
the size as 32bit value.