fix creation policy Scatter
It is possible that mallocMC is giving the same memory slot to different concurrent allocation calls.
The error shows up in very seldom cases in PIConGPU where many threads on CUDA Ampere GPUs allocated memory of different sizes at the same time.

https://github.com/ComputationalRadiationPhysics/picongpu/issues/4576#issuecomment-1632575759

The data race happens because the last thread is emptying the PTE with `_page[page].init();`. Most likely the data race happened because we alter the bitmasks with atomics but reset the bitmask in init without. A second potential reason could be that we read the chunk size in one place without using atomic which is resulting into data races too.

This PR
- removes reading chunk size without atomic
- is changing the way how a PTE is freed
- reduce branch divergence by refactoring the calls to `tryUsePage()`